requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
tokenizers>=0.13.0
Pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
tqdm>=4.65.0
wandb>=0.15.0
matplotlib>=3.7.0
seaborn>=0.12.0
requests>=2.31.0
psutil>=5.9.0
timm>=0.9.0
